requirements.txt
torch>=2.0.0
transformers>=4.35.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
httpx>=0.25.0
click>=8.1.0
python-dotenv>=1.0.0
Pillow>=9.0.0
scikit-learn>=1.0.0